FEniCS backend: Move interpolation code
